docketeer/pyproject.toml
[project]
name = "docketeer"
dynamic = ["version"]
description = "A reasonably sized autonomous AI construction kit"
readme = "README.md"
authors = [
    { name = "Chris Guidry", email = "guid@omg.lol" }
]
license = "MIT"
requires-python = ">=3.12"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
]
dependencies = [
    "croniter",
    "mcp",
    "pydocket>=0.17.8",
    "pyyaml",
    "watchfiles",
]

[project.urls]
Homepage = "https://github.com/chrisguidry/docketeer"
Repository = "https://github.com/chrisguidry/docketeer"
Issues = "https://github.com/chrisguidry/docketeer/issues"

[project.scripts]
docketeer = "docketeer.main:run"

[project.entry-points."docketeer.prompt"]
core = "docketeer.prompt:core_prompt"

[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "vcs"
raw-options.root = ".."

[tool.hatch.build.targets.wheel]
packages = ["src/docketeer"]

[tool.docketeer.ci]
apt-packages = ["bubblewrap", "socat"]
sysctl = ["kernel.apparmor_restrict_unprivileged_userns=0"]

[tool.pytest.ini_options]
minversion = "9.0"
timeout = 1
addopts = [
    "--import-mode=importlib",
    "--cov=docketeer",
    "--cov=tests",
    "--cov-config=../pyproject.toml",
    "--cov-branch",
    "--cov-report=term-missing",
    "--cov-fail-under=100",
]
asyncio_mode = "auto"
filterwarnings = [
    "error",
]
